Extract into Parser, Storage, TaskList, Ui
